requirements.txt
asgiref==3.2.10
Django==3.1
django-jsonfield==1.4.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
